refactor: move GTR counting helpers to common, fix root-state bug, delete gtr_inference_dense
- Move get_branch_mutation_matrix and accumulate_mutation_counts from ancestral/gtr_inference_dense.rs to gtr/infer_gtr/common.rs
- Convert count_transitions_from_marginal_data free fn to MarginalData::count_transitions method
- Fix root-state bug: iterate all profile rows instead of only row(0) (was correct for discrete/1-row, wrong for dense/multi-row)
- Move sparse counting helpers from module-level fns to PartitionMarginalSparse associated fns
- Delete dead code (infer_gtr_dense, get_mutation_counts_dense) and gtr_inference_dense.rs
- Rewrite test imports to use TransitionCounting trait + infer_gtr_impl
- Regenerate golden master outputs (branch-length clamping now consistent with marginal passes)
