+++ packages/treetime/src/partition/marginal_core.rs
@@ -1,7 +1,8 @@
-use crate::ancestral::gtr_inference_dense::{accumulate_mutation_counts, get_branch_mutation_matrix};
 use crate::constants::SUPERTINY_NUMBER;
 use crate::gtr::gtr::GTR;
-use crate::gtr::infer_gtr::common::{MutationCounts, is_profile_informative};
+use crate::gtr::infer_gtr::common::{
+  MutationCounts, accumulate_mutation_counts, get_branch_mutation_matrix, is_profile_informative,
+};
 use crate::partition::dense::{DenseEdgePartition, DenseNodePartition, DenseSeqDistribution, DenseSeqInfo};
 use eyre::Report;
 use itertools::izip;
@@ -21,12 +22,6 @@ pub struct MarginalData {
   pub nodes: BTreeMap<GraphNodeKey, DenseNodePartition>,
   pub edges: BTreeMap<GraphEdgeKey, DenseEdgePartition>,
   pub min_branch_length: f64,
-}
-
-impl MarginalData {
-  pub fn effective_branch_length(&self, raw: f64) -> f64 {
-    raw.max(self.min_branch_length)
-  }
 }
 
 pub trait MarginalPartition<N, E>: Send + Sync
@@ -193,46 +188,51 @@ pub fn normalize_inplace(dis: &mut Array2<f64>) -> f64 {
   log_lh
 }
 
-/// Count posterior-weighted transitions from `MarginalData`.
-///
-/// Shared by dense and discrete partitions (both store full profile matrices).
-pub fn count_transitions_from_marginal_data<N, E>(
-  data: &MarginalData,
-  graph: &Graph<N, E, ()>,
-) -> Result<MutationCounts, Report>
-where
-  N: GraphNode,
-  E: GraphEdge + HasBranchLength,
-{
-  let n_states = data.gtr.pi.len();
-  let mut nij = Array2::zeros((n_states, n_states));
-  let mut Ti = Array1::zeros(n_states);
-
-  for edge in graph.get_edges() {
-    let edge_arc = edge.read_arc();
-    let branch_length = data.effective_branch_length(edge_arc.payload().read_arc().branch_length().unwrap_or(0.0));
-    let edge_key = edge_arc.key();
-
-    let edge_data = &data.edges[&edge_key];
-
-    let exp_qt = data.gtr.expQt(branch_length) + SUPERTINY_NUMBER;
-    let mut_stack = get_branch_mutation_matrix(&edge_data.msg_to_child.dis, &edge_data.msg_to_parent.dis, &exp_qt);
-    accumulate_mutation_counts(&mut_stack, branch_length, &mut nij, &mut Ti);
+impl MarginalData {
+  pub fn effective_branch_length(&self, raw: f64) -> f64 {
+    raw.max(self.min_branch_length)
   }
 
-  let root = graph.get_exactly_one_root()?;
-  let root_key = root.read_arc().key();
-  let root_profile = data.nodes[&root_key].profile.dis.row(0);
-  let mut root_state = Array1::zeros(n_states);
-  if is_profile_informative(root_profile, n_states) {
-    if let Some(root_idx) = argmax_first(&root_profile) {
-      root_state[root_idx] = 1.0;
+  /// Count posterior-weighted transitions from dense profile matrices.
+  ///
+  /// Shared by dense and discrete partitions (both store full profile matrices).
+  pub fn count_transitions<N, E>(&self, graph: &Graph<N, E, ()>) -> Result<MutationCounts, Report>
+  where
+    N: GraphNode,
+    E: GraphEdge + HasBranchLength,
+  {
+    let n_states = self.gtr.pi.len();
+    let mut nij = Array2::zeros((n_states, n_states));
+    let mut Ti = Array1::zeros(n_states);
+
+    for edge in graph.get_edges() {
+      let edge_arc = edge.read_arc();
+      let branch_length = self.effective_branch_length(edge_arc.payload().read_arc().branch_length().unwrap_or(0.0));
+      let edge_key = edge_arc.key();
+
+      let edge_data = &self.edges[&edge_key];
+
+      let exp_qt = self.gtr.expQt(branch_length) + SUPERTINY_NUMBER;
+      let mut_stack = get_branch_mutation_matrix(&edge_data.msg_to_child.dis, &edge_data.msg_to_parent.dis, &exp_qt);
+      accumulate_mutation_counts(&mut_stack, branch_length, &mut nij, &mut Ti);
     }
+
+    let root = graph.get_exactly_one_root()?;
+    let root_key = root.read_arc().key();
+    let root_profile = &self.nodes[&root_key].profile.dis;
+    let mut root_state = Array1::zeros(n_states);
+    for row in root_profile.rows() {
+      if is_profile_informative(row, n_states) {
+        if let Some(root_idx) = argmax_first(&row) {
+          root_state[root_idx] += 1.0;
+        }
+      }
+    }
+
+    nij.diag_mut().fill(0.0);
+
+    Ok(MutationCounts { nij, Ti, root_state })
   }
-
-  nij.diag_mut().fill(0.0);
-
-  Ok(MutationCounts { nij, Ti, root_state })
 }
 
 pub fn normalize_from_log(log_dis: &Array2<f64>) -> (Array2<f64>, f64) {
